+++ packages/solid/CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to the library will be documented in this file.
 
 ## vX.X.X (Month DD, YYYY)
 
-- Change type of `value` validation function
-- Add support for validation of optional fields (issue #5)
 - Add support for scheme validation with Zod
+- Add support for validation of optional fields (issue #5)
+- Change type of `custom` and `value` validation function
 
 ## v0.8.4 (November 29, 2022)